% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_tree.R
\name{belong}
\alias{belong}
\title{Title: Provide each elements from setB belongs to set membership of setA}
\usage{
belong(setA, subB)
}
\arguments{
\item{setA}{whether the elements in this setA}

\item{subB}{the elements takes from subsetB}
}
\value{
A list of logic values indicates the membership;
the length of this list is the same as subB
}
\description{
\code{belong} provides whether each element of setB belong to setA
}
\details{
(TBF)
}
\examples{
library(tidyverse)
B <- c(2, 3, 4)
A <- 3:10
belong(setA = A, subB = B)
}
