% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_tree_structure.R
\name{Btree}
\alias{Btree}
\title{Title: Create the Coefficients for Each Interior Knots Matrix}
\usage{
Btree(Ytree, B)
}
\arguments{
\item{Ytree}{The results from \code{Ytree()}}

\item{B}{a list of covariate coefficients}
}
\value{
A set of n * 2 matrices, each of which represent the covariate coefficients
at the an interior knot and its children branches.
}
\description{
The Dirichlet Multinational Tree model requires the
outcomes at each interior knots. We need to change a list of coefficients
into high dimensional matrix format corresponding to each matrix in \code{Ytree} outcomes.
The length for \code{Btree} outcomes is the same as \code{Ytree}; TBF
}
\details{
This is adapted from \link{Tao Wang and Hongyu Zhao (2017)}
}
