% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_kfold.R
\name{cv_T_group_fun}
\alias{cv_T_group_fun}
\title{Title}
\usage{
cv_T_group_fun(
  Y,
  X,
  treeinfo,
  alpha = 0.5,
  cutoff = 0.8,
  model = "dirmult",
  err.conv = 0.001,
  iter.max = 30,
  L.init = NULL,
  lambda.max = NULL,
  lambda.min = NULL,
  nfolds = 10,
  CV = T,
  cv.ind = NULL
)
}
\arguments{
\item{cv.ind}{}
}
\value{

}
\description{
Title
}
