% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_likelihood.R
\name{f_fun}
\alias{f_fun}
\title{Title Penalized log-likelihood for Dirichlet Tree Multinomial Regression}
\usage{
f_fun(Ytree, X, B, model, alpha, lambda)
}
\arguments{
\item{lambda}{}
}
\value{

}
\description{
Title Penalized log-likelihood for Dirichlet Tree Multinomial Regression
}
