% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_kfold.R
\name{kfold.dismo}
\alias{kfold.dismo}
\title{Title}
\usage{
kfold.dismo(x, k = 5, by = NULL)
}
\arguments{
\item{by}{}
}
\value{

}
\description{
Title
}
\details{
adapted from R package dismo
??dismo # Species Distribution Modeling
}
