% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_dirichlet.R
\name{lambda_raw_fun}
\alias{lambda_raw_fun}
\title{Title Grid methods to find which lambda minimize the loss}
\usage{
lambda_raw_fun(grad, L, alpha, lambda.raw = 2, fac1 = 1.1, fac2 = 0.96)
}
\arguments{
\item{fac2}{}
}
\value{

}
\description{
Title Grid methods to find which lambda minimize the loss
}
