% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_likelihood.R
\name{Loglik_Dirichlet_tree}
\alias{Loglik_Dirichlet_tree}
\title{Title Log likelihood for Dirichlet Tree}
\usage{
Loglik_Dirichlet_tree(Ytree, X, B, model)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title Log likelihood for Dirichlet Tree
}
