% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_likelihood.R
\name{Score}
\alias{Score}
\title{Title Score function for Dirichlet and Multinomial models}
\usage{
Score(Y, X, b, model)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title Score function for Dirichlet and Multinomial models
}
