% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_simulation.R
\name{simulate_correlated_DTM}
\alias{simulate_correlated_DTM}
\title{Simulate Dirichlet-Multinomial-Tree Longitudinal Regression Model}
\usage{
simulate_correlated_DTM(
  subject_sim = 100,
  n_measurement = 3,
  X = NULL,
  zeta_sim = NULL,
  tree = NULL,
  num_leaf = 10,
  covariates_sim = 50,
  rho = NULL,
  Sigma = NULL,
  global_random = TRUE,
  random_time = FALSE,
  random_sim_mean = NULL,
  random_sim_cov = NULL,
  random_sim_mean_local = NULL,
  random_sim_cov_local = NULL,
  rep = rep,
  num_branch = 3,
  num_cov = 5,
  phi_min = 0.9,
  phi_max = 1.2,
  seed = 555
)
}
\arguments{
\item{subject_sim}{\code{integer} the number of subjects to simulate, default = 100}

\item{n_measurement}{\code{integer} number of measurements for each subject, default = 3}

\item{X}{design matrix of covariate}

\item{zeta_sim}{true inclusion indicators (must be provided if X is imported in argument)}

\item{tree}{\code{phylo} a tree structure for simulated data}

\item{num_leaf}{\code{integer} the number of leaves in tree to simulate, default = 5}

\item{covariates_sim}{\code{integer} the number of covariates to simulate, default = 50}

\item{rho}{\code{numeric} the correlation structure, default = 0.20}

\item{Sigma}{\code{matrix} the variance-covariance matrix must be symmetric positive semi-definite}

\item{global_random}{\code{logi} whether all the taxa share the same random effect or not, default = TRUE}

\item{random_time}{\code{logi} whether to include random slope term for time or not, default = FALSE}

\item{random_sim_mean}{a vector of mean used to simulate random effect, length = 1 if random intercept only}

\item{random_sim_cov}{a covariance matrix used to simulate random effect, length = 1 if random intercept only}

\item{random_sim_mean_local}{a vector of mean if local random effect is in used, length = number of taxa}

\item{random_sim_cov_local}{a covariance matrix if local random effect is in used, length = number of taxa}

\item{num_branch}{\code{integer} number of branches the associated covariates are in, default = 3}

\item{num_cov}{\code{integer} the number of associated covariates to simulate, default = 5}

\item{seed}{\code{integer} seed setup, default = 555}
}
\value{
A list of results:
\itemize{
\item Y: the compositional counts
\item X: the covariates design matrix
\item phi_sim: true regression coefficients
\item zeta_sim: true inclusion indicators
\item tree: the random tree, \code{phylo} object generated by the \code{ape} package
\item Sigma: the simulated covariance matrix
}
}
\description{
This is the adaptation of previous simulation function to the
longitudinal setting, where we can have repeated measure for each of the individual
}
\examples{
# random intercept only
data_random_intercept_global <- simulate_correlated_DTM(rho = 0.2, random_sim_mean = 0, random_sim_cov = 1, rep = 100)
# random intercept and slope (uncorrelated)
data_random_intercept_slope_global <- simulate_correlated_DTM(rho = 0.2, X = data_random_intercept_global$X, zeta_sim = data_random_intercept_global$zeta_sim, random_sim_mean = c(1,1), random_sim_cov = diag(1,2), rep = 100, random_time = TRUE)
# random intercept and random slope (correlated)

data_random_intercept_slope_correlated_global <- simulate_correlated_DTM(rho = 0.2, X = data_random_intercept_global$X, zeta_sim = data_random_intercept_global$zeta_sim,random_sim_mean = c(1,1), random_sim_cov =  matrix(c(1,0.2,0.2,1), nrow = 2), rep = 100, random_time = TRUE, seed = 100)

}
