% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_simulation.R
\name{simulate_DM}
\alias{simulate_DM}
\title{Title Simulate Dirichlet-Multinomial Regression Model}
\usage{
simulate_DM(
  n_obs = 100,
  n_vars = 30,
  n_taxa = 75,
  n_relevant_vars = 4,
  n_relevant_taxa = 4,
  beta_min = 1,
  beta_max = 1.25,
  signoise = 1,
  n_reads_min = 5000,
  n_reads_max = 10000,
  theta0 = 0.01,
  rho = NULL,
  Sigma = NULL
)
}
\arguments{
\item{Sigma}{}
}
\value{
A list of outcomes including:
\itemize{
\item X = XX
\item Y = YY
\item alphas = intercept,
\item betas = betas,
\item n_reads_min = n_reads_min,
\item n_reads_max = n_reads_max,
\item theta0 = theta0,
\item phi = phi,
\item rho = rho,
\item signoise = signoise,
\item Sigma = Sigma
}
}
\description{
Title Simulate Dirichlet-Multinomial Regression Model
}
\section{This function can be used to simulate DM data.}{
Taken from Wadsworth (2017)
An integrative Bayesian Dirichlet-multinomial regression model for
the analysis of taxonomic abundances in microbiome data
}

