% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_simulation.R
\name{simulate_DMLM}
\alias{simulate_DMLM}
\title{Title simulate data for DMLMbvs model}
\usage{
simulate_DMLM(
  subject_sim = 50,
  B_sim = 50,
  n_vars = 50,
  active_cov = 10,
  rho = NULL,
  Sigma = NULL,
  seed = 111
)
}
\arguments{
\item{seed}{}
}
\value{
a list of outcomes:
\itemize{
\item Y = Y,
\item Z = Z,
\item X = X,
\item true_cov = true_cov,
\item true_coeff = true_coeff,
\item true_coeff_beta = true_coeff_beta)
}
}
\description{
Title simulate data for DMLMbvs model
}
