% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{soft_fun}
\alias{soft_fun}
\title{Title soft activation function}
\usage{
soft_fun(M, v)
}
\arguments{
\item{v}{}
}
\value{

}
\description{
Title soft activation function
}
\details{
this is the equation
}
