% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_dirichlet.R
\name{T_group_path}
\alias{T_group_path}
\title{Title to get the tree path and alpha for penalty likelihood}
\usage{
T_group_path(
  Y,
  X,
  treeinfo,
  alpha = 0.5,
  cutoff = 0.8,
  model = "dirmult",
  err.conv = 0.001,
  iter.max = 30,
  L.init = NULL,
  lambda.max = NULL,
  lambda.min = NULL
)
}
\arguments{
\item{lambda.min}{}
}
\value{

}
\description{
Title to get the tree path and alpha for penalty likelihood
}
