% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_tree.R
\name{treeinfo}
\alias{treeinfo}
\title{Extract Structural Information from a Phylo Tree}
\usage{
treeinfo(tree)
}
\arguments{
\item{tree}{A "phylo" structured tree file}
}
\value{
A list object contains necessary information from the \code{tree}.
The output has the following properties:
\itemize{
\item Nnodes: \code{integer} the number of nodes in the given "phylo" tree
\item Ntips: \code{integer} the number of tips/leaves in the given "phylo" tree
\item Tredge: \code{matrix} the matrix for the edges information;
each row represents the two ends of the edges
\item TreeMat: \code{matrix} the matrix consists of logical values;
the binary value of each taxa (by row) belongs to
a given leaf or inner node (by col).
}
}
\description{
\code{treeinfo()} provides the basic information from a phylo tree
without extract the distance.
}
\details{
This is adapted from \link{Tao Wang and Hongyu Zhao (2017)}
}
\examples{
library(ape)
library(caper)
tree <- rtree(n = 10)
plot(tree, edge.width = 2)

treeinfo(tree)

}
