% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_simulation.R
\name{Xsim}
\alias{Xsim}
\title{Title: Simulate Design matrix and Indicators}
\usage{
Xsim(
  subject_sim = 100,
  tree = NULL,
  num_leaf = 10,
  covariates_sim = 50,
  rho = NULL,
  Sigma = NULL,
  num_branch = 3,
  num_cov = 5,
  seed = 555
)
}
\arguments{
\item{subject_sim}{\code{integer} Number of subjects to simulate}

\item{tree}{\code{phylo} The function is set to simulate the tree structure for the
simulated data. Default is NULL.}

\item{covariates_sim}{\code{integer} Number of covariates to simulate}

\item{rho}{\code{numeric} Used for covariate correlation between 0 & 1. Default is NULL.
Need either rho or Sigma to be input.}

\item{Sigma}{\code{matrix} A symmetric positive definite matrix used for covariate
correlation structure columns must match number of covariates.
Default is NULL. Need either rho or Sigma to be input.}

\item{num_branch}{\code{integer} Number of tree branches}

\item{num_cov}{\code{integer} Number of associated covariates to simulate}

\item{seed}{\code{integer} The desired seed for reproducibility. Default is 555.}

\item{num_leaf`integer`}{Number of nodes in tree to simulate}
}
\value{
A list of outcomes including:
\itemize{
\item X = The design matrix
\item Zeta = The matrix indicating which covariates are significant
\item Sigma = The simulated covariance matrix
\item tree = The simulated tree, \code{phylo} object generated by the \code{ape} package
}
}
\description{
Title: Simulate Design matrix and Indicators
}
