% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_likelihood.R
\name{Hessian}
\alias{Hessian}
\title{Title Hessian matrix for Dirichlet and Multinomial models}
\usage{
Hessian(Y, X, b, model)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title Hessian matrix for Dirichlet and Multinomial models
}
