% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_dirichlet.R
\name{lambda_fun}
\alias{lambda_fun}
\title{Title Getting betas use the max(beta, 1)}
\usage{
lambda_fun(grad, L, alpha, lambda)
}
\arguments{
\item{lambda}{}
}
\value{

}
\description{
Title Getting betas use the max(beta, 1)
}
