% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_likelihood.R
\name{Loglik}
\alias{Loglik}
\title{Title: Likelihood function for Dirichlet and Multinomial models}
\usage{
Loglik(Y, X, b, model)
}
\arguments{
\item{Y}{= the vector of count outcomes}

\item{X}{\code{matrix} the matrix of covariates}

\item{b}{}

\item{model}{}
}
\value{

}
\description{
Title: Likelihood function for Dirichlet and Multinomial models
}
