% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_likelihood.R
\name{Loglik_Dirichlet_tree}
\alias{Loglik_Dirichlet_tree}
\title{Title: Log likelihood for Dirichlet Tree}
\usage{
Loglik_Dirichlet_tree(Ytree, X, B, model)
}
\arguments{
\item{Ytree}{is the tree information from the \code{Ytree} function.
Input will be a set of n * 2 matrices, each of which represent the an interior knot
and its children branches}

\item{X}{\code{matrix} of nxp  which is the number of subjects by number of covariates}

\item{B}{a list of covariate coefficients}

\item{model}{\code{character} type of model to use for the Log Likelihood. Options are
(Dirichlet Multinomial = "dirmult", Multinomial = "mult", or
Dirichlet = "dir")}
}
\value{
Will return the log likelihood for the Dirichlet Tree model for each of the nodes
}
\description{
Title: Log likelihood for Dirichlet Tree
}
\details{
Calculating the log likelihood for the Dirichlet tree
}
