% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_dirichlet.R
\name{QL_fun}
\alias{QL_fun}
\title{Title  Gradient Descent Method for penalized likelihood
this is gradient descent methods for
penalized likelihood}
\usage{
QL_fun(Ytree, X, W, model, B1, grad, alpha, lambda, L)
}
\arguments{
\item{L}{}
}
\value{

}
\description{
Title  Gradient Descent Method for penalized likelihood
this is gradient descent methods for
penalized likelihood
}
