% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_simulation.R
\name{simulate_DM}
\alias{simulate_DM}
\title{Title: Simulate Dirichlet-Multinomial Regression Model}
\usage{
simulate_DM(
  n_obs = 100,
  n_vars = 30,
  n_taxa = 75,
  n_relevant_vars = 4,
  n_relevant_taxa = 4,
  beta_min = 1,
  beta_max = 1.25,
  signoise = 1,
  n_reads_min = 5000,
  n_reads_max = 10000,
  theta0 = 0.01,
  rho = NULL,
  Sigma = NULL
)
}
\arguments{
\item{n_obs}{\code{integer} Number of observations}

\item{n_vars}{\code{integer} Number of variables}

\item{n_taxa}{\code{integer} Number of taxa}

\item{n_relevant_vars}{\code{integer} Number of variables that are important}

\item{n_relevant_taxa}{\code{integer} Number of important taxa}

\item{beta_min}{\code{numeric} The desired min beta value}

\item{beta_max}{\code{numeric} The desired max beta value}

\item{signoise}{\code{numeric} Default is 1.}

\item{n_reads_min}{\code{numeric} Default is 5000. Used to obtain a random subject from n_reads_min
to n_reads_max from the Dirichlet Multinomial Distribution using dirmult::simPop()}

\item{n_reads_max}{\code{numeric} Default is 10000. Used to obtain a random subject from n_reads_min
to n_reads_max from the Dirichlet Multinomial Distribution using dirmult::simPop()}

\item{rho}{\code{numeric} Used for covariate correlation between 0 & 1. Default is NULL.
Need either rho or Sigma to be input.}

\item{Sigma}{\code{matrix} A symmetric positive definite matrix used for covariate
correlation structure columns must match number of covariates.
Default is NULL. Need either rho or Sigma to be input.}
}
\value{
A list of outcomes including:
\itemize{
\item X = XX
\item Y = YY
\item alphas = intercept,
\item betas = betas,
\item n_reads_min = n_reads_min,
\item n_reads_max = n_reads_max,
\item theta0 = theta0,
\item phi = phi,
\item rho = rho,
\item signoise = signoise,
\item Sigma = Sigma
}
}
\description{
Title: Simulate Dirichlet-Multinomial Regression Model
}
\section{This function can be used to simulate DM data.}{
Taken from Wadsworth (2017)
An integrative Bayesian Dirichlet-Multinomial regression model for
the analysis of taxonomic abundances in microbiome data
}

