% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_simulation.R
\name{simulate_DMLM}
\alias{simulate_DMLM}
\title{Title simulate data for DMLMbvs model}
\usage{
simulate_DMLM(
  subject_sim = 50,
  B_sim = 50,
  n_vars = 50,
  active_cov = 10,
  rho = NULL,
  Sigma = NULL,
  seed = 111
)
}
\arguments{
\item{subject_sim}{\code{integer} Number of subjects to simulate}

\item{B_sim}{\code{integer} Number of betas}

\item{n_vars}{\code{integer} Number of covariates to simulate}

\item{active_cov}{\code{integer} Number of associated covariates}

\item{rho}{\code{numeric} Used for covariate correlation between 0 & 1. Default is NULL.
Need either rho or Sigma to be input.}

\item{Sigma}{\code{matrix} A symmetric positive definite matrix used for covariate
correlation structure columns must match number of covariates.
Default is NULL. Need either rho or Sigma to be input.}

\item{seed}{\code{integer} Seed for reproducibility, default = 555}
}
\value{
a list of outcomes:
\itemize{
\item Y = Y,
\item Z = Z,
\item X = X,
\item true_cov = true_cov,
\item true_coeff = true_coeff,
\item true_coeff_beta = true_coeff_beta)
}
}
\description{
Title simulate data for DMLMbvs model
}
