% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_simulation.R
\name{simulate_DTM}
\alias{simulate_DTM}
\title{Title Simulate Dirichlet-Multinomial-Tree Regression Model}
\usage{
simulate_DTM(
  subject_sim = 100,
  tree = NULL,
  num_leaf = 10,
  covariates_sim = 50,
  rho = NULL,
  Sigma = NULL,
  num_branch = 3,
  num_cov = 5,
  phi_min = 0.9,
  phi_max = 1.2,
  seed = 555
)
}
\arguments{
\item{subject_sim}{\code{integer} the number of subjects to simulate, default = 100}

\item{tree}{\code{phylo} a tree structure for simulated data}

\item{num_leaf}{\code{integer} the number of leaves in tree to simulate, default = 5}

\item{covariates_sim}{\code{integer} the number of covariates to simulate, default = 50}

\item{rho}{\code{numeric} the correlation structure, default = 0.20}

\item{Sigma}{\code{matrix} the variance-covariance matrix must be symmetric positive semi-definite}

\item{num_branch}{\code{integer} number of branches the associated covariates are in, default = 3}

\item{num_cov}{\code{integer} the number of associated covariates to simulate, default = 5}

\item{seed}{\code{integer} seed setup, default = 555}
}
\value{
A list of results:
\itemize{
\item Y: the compositional counts
\item X: the covariates design matrix
\item phi_sim: true regression coefficients
\item zeta_sim: true inclusion indicators
\item tree: the random tree, \code{phylo} object generated by the \code{ape} package
\item Sigma: the simulated covariance matrix
}
}
\description{
Wrapper function for the Rcpp code to simulate DTM data
Taken from Wadsworth (2017)
}
