% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_simulation.R
\name{Xsim}
\alias{Xsim}
\title{Title Simulate Design matrix and Indicators}
\usage{
Xsim(
  subject_sim = 100,
  tree = NULL,
  num_leaf = 10,
  covariates_sim = 50,
  rho = NULL,
  Sigma = NULL,
  num_branch = 3,
  num_cov = 5,
  seed = 555
)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Title Simulate Design matrix and Indicators
}
